sr(api): move find-paginated-target-profile-training-summaries in devcomp
