+++ api/tests/devcomp/unit/domain/usecases/find-paginated-target-profile-training-summaries_test.js
@@ -1,5 +1,5 @@
-import { sinon, expect } from '../../../test-helper.js';
-import { findPaginatedTargetProfileTrainingSummaries as findPaginatedTargetProfileTrainings } from '../../../../lib/domain/usecases/find-paginated-target-profile-training-summaries.js';
+import { sinon, expect } from '../../../../test-helper.js';
+import { findPaginatedTargetProfileTrainingSummaries as findPaginatedTargetProfileTrainings } from '../../../../../src/devcomp/domain/usecases/find-paginated-target-profile-training-summaries.js';
 
 describe('Unit | UseCase | findPaginatedTargetProfileTrainingSummaries', function () {
   it('should call the repository with the right arguments and return summaries', async function () {
